bump ext-relay to 0.12+
