+++ src/Symfony/Component/Cache/Traits/RelayProxy.php
@@ -11,7 +11,6 @@
 
 namespace Symfony\Component\Cache\Traits;
 
-use Symfony\Component\Cache\Traits\Relay\Relay12Trait;
 use Symfony\Component\VarExporter\LazyObjectInterface;
 use Symfony\Contracts\Service\ResetInterface;
 
@@ -28,7 +27,6 @@ class RelayProxy extends \Relay\Relay implements ResetInterface, LazyObjectInter
     use RedisProxyTrait {
         resetLazyObject as reset;
     }
-    use Relay12Trait;
 
     public function __construct($host = null, $port = 6379, $connect_timeout = 0.0, $command_timeout = 0.0, #[\SensitiveParameter] $context = [], $database = 0)
     {
@@ -268,6 +266,11 @@ class RelayProxy extends \Relay\Relay implements ResetInterface, LazyObjectInter
     public function del(...$keys): \Relay\Relay|bool|int
     {
         return $this->initializeLazyObject()->del(...\func_get_args());
+    }
+
+    public function delifeq($key, $value): \Relay\Relay|false|int
+    {
+        return $this->initializeLazyObject()->delifeq(...\func_get_args());
     }
 
     public function discard(): bool
@@ -1420,6 +1423,71 @@ class RelayProxy extends \Relay\Relay implements ResetInterface, LazyObjectInter
         return $this->initializeLazyObject()->unwatch(...\func_get_args());
     }
 
+    public function vadd($key, $values, $element, $options = null): \Relay\Relay|false|int
+    {
+        return $this->initializeLazyObject()->vadd(...\func_get_args());
+    }
+
+    public function vcard($key): \Relay\Relay|false|int
+    {
+        return $this->initializeLazyObject()->vcard(...\func_get_args());
+    }
+
+    public function vdim($key): \Relay\Relay|false|int
+    {
+        return $this->initializeLazyObject()->vdim(...\func_get_args());
+    }
+
+    public function vemb($key, $element, $raw = false): \Relay\Relay|array|false
+    {
+        return $this->initializeLazyObject()->vemb(...\func_get_args());
+    }
+
+    public function vgetattr($key, $element, $raw = false): \Relay\Relay|array|false|string
+    {
+        return $this->initializeLazyObject()->vgetattr(...\func_get_args());
+    }
+
+    public function vinfo($key): \Relay\Relay|array|false
+    {
+        return $this->initializeLazyObject()->vinfo(...\func_get_args());
+    }
+
+    public function vismember($key, $element): \Relay\Relay|bool
+    {
+        return $this->initializeLazyObject()->vismember(...\func_get_args());
+    }
+
+    public function vlinks($key, $element, $withscores): \Relay\Relay|array|false
+    {
+        return $this->initializeLazyObject()->vlinks(...\func_get_args());
+    }
+
+    public function vrandmember($key, $count = 0): \Relay\Relay|array|false|string
+    {
+        return $this->initializeLazyObject()->vrandmember(...\func_get_args());
+    }
+
+    public function vrange($key, $min, $max, $count = -1): \Relay\Relay|array|false
+    {
+        return $this->initializeLazyObject()->vrange(...\func_get_args());
+    }
+
+    public function vrem($key, $element): \Relay\Relay|false|int
+    {
+        return $this->initializeLazyObject()->vrem(...\func_get_args());
+    }
+
+    public function vsetattr($key, $element, $attributes): \Relay\Relay|false|int
+    {
+        return $this->initializeLazyObject()->vsetattr(...\func_get_args());
+    }
+
+    public function vsim($key, $member, $options = null): \Relay\Relay|array|false
+    {
+        return $this->initializeLazyObject()->vsim(...\func_get_args());
+    }
+
     public function wait($replicas, $timeout): \Relay\Relay|false|int
     {
         return $this->initializeLazyObject()->wait(...\func_get_args());
@@ -1440,6 +1508,11 @@ class RelayProxy extends \Relay\Relay implements ResetInterface, LazyObjectInter
         return $this->initializeLazyObject()->xack(...\func_get_args());
     }
 
+    public function xackdel($key, $group, $ids, $mode = null): \Relay\Relay|array|false
+    {
+        return $this->initializeLazyObject()->xackdel(...\func_get_args());
+    }
+
     public function xadd($key, $id, $values, $maxlen = 0, $approx = false, $nomkstream = false): \Relay\Relay|false|null|string
     {
         return $this->initializeLazyObject()->xadd(...\func_get_args());
@@ -1458,6 +1531,11 @@ class RelayProxy extends \Relay\Relay implements ResetInterface, LazyObjectInter
     public function xdel($key, $ids): \Relay\Relay|false|int
     {
         return $this->initializeLazyObject()->xdel(...\func_get_args());
+    }
+
+    public function xdelex($key, $ids, $mode = null): \Relay\Relay|array|false
+    {
+        return $this->initializeLazyObject()->xdelex(...\func_get_args());
     }
 
     public function xgroup($operation, $key = null, $group = null, $id_or_consumer = null, $mkstream = false, $entries_read = -2): mixed
